+++ packages/horizon/test/integration/during-transition-period/operator.test.ts
@@ -1,0 +1,93 @@
+import hre from 'hardhat'
+
+import { ethers } from 'hardhat'
+import { expect } from 'chai'
+import { generatePOI } from '@graphprotocol/toolshed'
+import { getEventData } from '@graphprotocol/toolshed/hardhat'
+import { indexers } from '../../../tasks/test/fixtures/indexers'
+
+import type { HardhatEthersSigner } from '@nomicfoundation/hardhat-ethers/signers'
+import type { HorizonStakingExtension } from '@graphprotocol/toolshed/deployments'
+
+describe('Operator', () => {
+  let snapshotId: string
+
+  // Subgraph service address is not set for integration tests
+  const subgraphServiceAddress = '0x0000000000000000000000000000000000000000'
+
+  const graph = hre.graph()
+  const horizonStaking = graph.horizon.contracts.HorizonStaking
+
+  beforeEach(async () => {
+    // Take a snapshot before each test
+    snapshotId = await ethers.provider.send('evm_snapshot', [])
+  })
+
+  afterEach(async () => {
+    // Revert to the snapshot after each test
+    await ethers.provider.send('evm_revert', [snapshotId])
+  })
+
+  describe('Existing Protocol Users', () => {
+    let indexer: HardhatEthersSigner
+    let operator: HardhatEthersSigner
+    let allocationID: string
+    let allocationTokens: bigint
+    let delegationIndexingCut: number
+
+    before(async () => {
+      const indexerFixture = indexers[0]
+      const allocationFixture = indexerFixture.allocations[0]
+
+      // Get signers
+      indexer = await ethers.getSigner(indexerFixture.address)
+      ;[operator] = await graph.accounts.getTestAccounts()
+
+      // Get allocation details
+      allocationID = allocationFixture.allocationID
+      allocationTokens = allocationFixture.tokens
+      delegationIndexingCut = indexerFixture.indexingRewardCut
+
+      // Set the operator
+      await horizonStaking.connect(indexer).setOperator(subgraphServiceAddress, operator.address, true)
+    })
+
+    it('should allow the operator to close an open legacy allocation and collect rewards', async () => {
+      // Use a non-zero POI
+      const poi = generatePOI('poi')
+      const thawingPeriod = await horizonStaking.__DEPRECATED_getThawingPeriod()
+
+      // Get delegation pool before closing allocation
+      const delegationPoolBefore = await horizonStaking.getDelegationPool(indexer.address, subgraphServiceAddress)
+      const delegationPoolTokensBefore = delegationPoolBefore.tokens
+
+      // Mine blocks to simulate time passing
+      const halfThawingPeriod = Number(thawingPeriod) / 2
+      for (let i = 0; i < halfThawingPeriod; i++) {
+        await ethers.provider.send('evm_mine', [])
+      }
+
+      // Get idle stake before closing allocation
+      const idleStakeBefore = await horizonStaking.getIdleStake(indexer.address)
+
+      // Close allocation
+      const tx = await (horizonStaking as HorizonStakingExtension).connect(operator).closeAllocation(allocationID, poi)
+      const eventData = await getEventData(tx, 'event HorizonRewardsAssigned(address indexed indexer, address indexed allocationID, uint256 amount)')
+      const rewards = eventData[2]
+
+      // Verify rewards are not zero
+      expect(rewards).to.not.equal(0, 'Rewards were not transferred to service provider')
+
+      // Verify rewards minus delegation cut are restaked
+      const idleStakeAfter = await horizonStaking.getIdleStake(indexer.address)
+      const idleStakeRewardsTokens = rewards * BigInt(delegationIndexingCut) / 1000000n
+      expect(idleStakeAfter).to.equal(idleStakeBefore + allocationTokens + idleStakeRewardsTokens, 'Rewards were not restaked')
+
+      // Verify delegators cut is added to delegation pool
+      const delegationPool = await horizonStaking.getDelegationPool(indexer.address, subgraphServiceAddress)
+      const delegationPoolTokensAfter = delegationPool.tokens
+      const delegationRewardsTokens = rewards - idleStakeRewardsTokens
+      expect(delegationPoolTokensAfter).to.equal(delegationPoolTokensBefore + delegationRewardsTokens, 'Delegators cut was not added to delegation pool')
+    })
+  })
+})