+++ packages/toolshed/eslint.config.js
@@ -1,0 +1,3 @@
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const config = require('eslint-graph-config')
+module.exports = config.default